+++ src/collective/easyform/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 from zope.i18nmessageid import MessageFactory