+++ src/collective/easyform/api.py
@@ -1,7 +1,13 @@
 # -*- coding: utf-8 -*-
+from AccessControl import ClassSecurityInfo
+from App.class_init import InitializeClass
+from collections import OrderedDict as BaseDict
 from collective.easyform.config import MODEL_DEFAULT
+from collective.easyform.interfaces import IFieldExtender
 from email.utils import formataddr
 from plone import api
+from plone.namedfile.interfaces import INamedBlobFile
+from plone.namedfile.interfaces import INamedFile
 from plone.supermodel import loadString
 from plone.supermodel import serializeSchema
 from plone.supermodel.parser import SupermodelParseError
@@ -10,6 +16,7 @@ from Products.CMFCore.Expression import getExprContext
 from Products.CMFPlone.utils import safe_unicode
 from re import compile
 from types import StringTypes
+from zope.schema import getFieldsInOrder
 
 
 CONTEXT_KEY = u'context'
@@ -18,8 +25,24 @@ CONTEXT_KEY = u'context'
 dollarRE = compile(r'\$\{(.+?)\}')
 
 
-class DollarVarReplacer(object):
+class OrderedDict(BaseDict):
+    """
+    A wrapper around dictionary objects that provides an ordering for
+    keys() and items().
+    """
+    security = ClassSecurityInfo()
+    security.setDefaultAccess('allow')
 
+    def reverse(self):
+        items = list(self.items())
+        items.reverse()
+        return items
+
+
+InitializeClass(OrderedDict)
+
+
+class DollarVarReplacer(object):
     """
     Initialize with a dictionary, then self.sub returns a string
     with all ${key} substrings replaced with values looked
@@ -241,3 +264,53 @@ def lnbr(text):
     """Converts line breaks to html breaks
     """
     return "<br/>".join(text.strip().splitlines()) if text else text
+
+
+def is_file_data(field):
+    """Return True, if field is a file field.
+    """
+    ifaces = (INamedFile, INamedBlobFile)
+    for i in ifaces:
+        if i.providedBy(field):
+            return True
+    return False
+
+
+def filter_fields(context, schema, unsorted_data, omit=False):
+    """Filter according to ``showAll``, ``showFields`` and ``includeEmpties``
+    settings to display in result mailings, thanks pages and the like.
+    """
+    data = OrderedDict([
+        (x[0], unsorted_data[x[0]])
+        for x in getFieldsInOrder(schema)
+        if x[0] in unsorted_data
+    ])
+
+    # TODO: Exclude labels
+    fields = [
+        f for f in data
+        if not (is_file_data(data[f])) and not (
+            # TODO: serverSide should always be excluded
+            # Also, when shoAll = 0 and serverSite = 1, it will be included
+            # which is wrong.
+            getattr(context, 'showAll', True) and
+            IFieldExtender(schema[f]).serverSide
+        )
+    ]
+
+    if not getattr(context, 'showAll', True):
+        showFields = getattr(context, 'showFields', []) or []
+        fields = [f for f in fields if f in showFields]
+
+    if not getattr(context, 'includeEmpties', True):
+        fields = [f for f in fields if data[f]]
+
+    ret = []
+    if omit:
+        # return field ids to omit
+        ret = [f for f in data if f not in fields]
+    else:
+        # return field ids and field instances to include
+        ret = OrderedDict([(f, data[f]) for f in fields])
+
+    return ret